+++ hklpy2/diffract.py
@@ -15,10 +15,12 @@ from collections.abc import Iterable
 from typing import Callable
 from typing import Optional
 
+import numpy as np
 import yaml
 from ophyd import Component as Cpt
 from ophyd import EpicsMotor
 from ophyd import Kind
+from ophyd import PositionerBase
 from ophyd import PseudoPositioner
 from ophyd import PseudoSingle
 from ophyd import SoftPositioner
@@ -612,11 +614,33 @@ class DiffractometerBase(PseudoPositioner):
     def wh(self, digits=4, full=False):
         """Concise report of the current diffractometer positions."""
 
-        def wh_round(label, value):
+        def labeled_value(label, value):
             return f"{label}={roundoff(value, digits)}"
 
-        def print_axes(names):
-            print(", ".join([wh_round(nm, getattr(self, nm).position) for nm in names]))
+        def print_axes(names: list[str], preface: str = ""):
+            if len(names):
+                value_text = ", ".join(
+                    [
+                        # Any instance of PositionerBase.
+                        labeled_value(nm, getattr(self, nm).position)
+                        for nm in names
+                    ]
+                )
+                print(preface + value_text)
+
+        def format_array(array):
+            """Apply (fixed-point) digits formatting to numpy array."""
+
+            def each(x):
+                """Roundoff each float."""
+                # return roundoff(x, digits)
+                text = f"{x:.{digits}f}"
+                if text.startswith("-0."):
+                    text = text[1:]
+                return float(text)
+
+            # Apply formatting function to each element.
+            return np.vectorize(each)(np.array(array)).tolist()
 
         if full:
             print(f"diffractometer={self.name!r}")
@@ -625,20 +649,35 @@ class DiffractometerBase(PseudoPositioner):
             for v in self.sample.reflections.values():
                 print(f"{v}")
             print(f"Orienting reflections: {self.sample.reflections.order}")
-            print(f"U={self.sample.U}")
-            print(f"UB={self.sample.UB}")
+            print(f"U={format_array(self.sample.U)}")
+            print(f"UB={format_array(self.sample.UB)}")
             for v in self.core.constraints.values():
                 print(f"constraint: {v}")
             print(f"Mode: {self.core.mode}")
-            print(f"beam={self.beam._asdict()}")
+            beam = self.beam._asdict()
+            for key in "energy wavelength".split():
+                if key in beam:
+                    beam[key] = roundoff(beam[key], digits)
+            print(f"beam={beam}")
         else:
-            print(f"wavelength={self.beam.wavelength.get()}")
+            print(f"wavelength={roundoff(self.beam.wavelength.get(), digits)}")
 
-        print_axes(self.pseudo_axis_names)
-        print_axes(self.real_axis_names)
+        print_axes(self.pseudo_axis_names, preface="pseudos: ")
+        print_axes(self.real_axis_names, preface="reals: ")
         extras = self.core.extras
         if len(extras) > 0:
-            print(" ".join([wh_round(k, v) for k, v in extras.items()]))
+            value_text = " ".join([labeled_value(k, v) for k, v in extras.items()])
+            print("extras: " + value_text)
+        print_axes(
+            [
+                # Names of any positioners not described above.
+                attr
+                for attr in self.component_names
+                if isinstance(getattr(self, attr), PositionerBase)
+                if attr not in (self.pseudo_axis_names + self.real_axis_names)
+            ],
+            preface="positioners: ",
+        )
 
 
 def creator(